python_ml/requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
requests>=2.31.0
beautifulsoup4>=4.12.2
pymongo>=4.6.0
python-dotenv>=1.0.0
schedule>=1.2.0
lxml>=5.1.0
joblib>=1.3.2
